fix(pages): isolate script loader bootstrap
